split out initial-runner
